+++ docs/changes.md
@@ -2,6 +2,7 @@
 
 2025/08/28
   * 完善tree_node_feature_segment(感谢兆坤提供补丁)
+  * 统一foreach visit返回值处理逻辑(感谢兆坤提供补丁)
 
 2025/08/23
   * 完善 dlist_remove_ex